server/requirement.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-cors-headers==4.3.1
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.0.7
